Split StreamEnconder and Stream mocks out
More work to minimize the size of headers in UTs to hopefully speed them
up.
