+++ test/mocks/http/mocks.cc
@@ -28,30 +28,6 @@ MockServerConnectionCallbacks::~MockServerConnectionCallbacks() {}
 MockStreamCallbacks::MockStreamCallbacks() {}
 MockStreamCallbacks::~MockStreamCallbacks() {}
 
-MockStream::MockStream() {
-  ON_CALL(*this, addCallbacks(_)).WillByDefault(Invoke([this](StreamCallbacks& callbacks) -> void {
-    callbacks_.push_back(&callbacks);
-  }));
-
-  ON_CALL(*this, removeCallbacks(_))
-      .WillByDefault(
-          Invoke([this](StreamCallbacks& callbacks) -> void { callbacks_.remove(&callbacks); }));
-
-  ON_CALL(*this, resetStream(_)).WillByDefault(Invoke([this](StreamResetReason reason) -> void {
-    for (StreamCallbacks* callbacks : callbacks_) {
-      callbacks->onResetStream(reason);
-    }
-  }));
-}
-
-MockStream::~MockStream() {}
-
-MockStreamEncoder::MockStreamEncoder() {
-  ON_CALL(*this, getStream()).WillByDefault(ReturnRef(stream_));
-}
-
-MockStreamEncoder::~MockStreamEncoder() {}
-
 MockServerConnection::MockServerConnection() {
   ON_CALL(*this, protocol()).WillByDefault(Return(protocol_));
 }
