+++ test/mocks/http/mocks.h
@@ -20,7 +20,9 @@
 #include "test/mocks/common.h"
 #include "test/mocks/event/mocks.h"
 #include "test/mocks/http/conn_pool.h"
+#include "test/mocks/http/stream.h"
 #include "test/mocks/http/stream_decoder.h"
+#include "test/mocks/http/stream_encoder.h"
 #include "test/mocks/router/mocks.h"
 #include "test/mocks/stream_info/mocks.h"
 #include "test/mocks/tracing/mocks.h"
@@ -64,50 +66,6 @@ public:
   MOCK_METHOD1(onResetStream, void(StreamResetReason reason));
   MOCK_METHOD0(onAboveWriteBufferHighWatermark, void());
   MOCK_METHOD0(onBelowWriteBufferLowWatermark, void());
-};
-
-class MockStream : public Stream {
-public:
-  MockStream();
-  ~MockStream();
-
-  // Http::Stream
-  MOCK_METHOD1(addCallbacks, void(StreamCallbacks& callbacks));
-  MOCK_METHOD1(removeCallbacks, void(StreamCallbacks& callbacks));
-  MOCK_METHOD1(resetStream, void(StreamResetReason reason));
-  MOCK_METHOD1(readDisable, void(bool disable));
-  MOCK_METHOD2(setWriteBufferWatermarks, void(uint32_t, uint32_t));
-  MOCK_METHOD0(bufferLimit, uint32_t());
-
-  std::list<StreamCallbacks*> callbacks_{};
-
-  void runHighWatermarkCallbacks() {
-    for (auto* callback : callbacks_) {
-      callback->onAboveWriteBufferHighWatermark();
-    }
-  }
-
-  void runLowWatermarkCallbacks() {
-    for (auto* callback : callbacks_) {
-      callback->onBelowWriteBufferLowWatermark();
-    }
-  }
-};
-
-class MockStreamEncoder : public StreamEncoder {
-public:
-  MockStreamEncoder();
-  ~MockStreamEncoder();
-
-  // Http::StreamEncoder
-  MOCK_METHOD1(encode100ContinueHeaders, void(const HeaderMap& headers));
-  MOCK_METHOD2(encodeHeaders, void(const HeaderMap& headers, bool end_stream));
-  MOCK_METHOD2(encodeData, void(Buffer::Instance& data, bool end_stream));
-  MOCK_METHOD1(encodeTrailers, void(const HeaderMap& trailers));
-  MOCK_METHOD1(encodeMetadata, void(const MetadataMap& metadata_map));
-  MOCK_METHOD0(getStream, Stream&());
-
-  testing::NiceMock<MockStream> stream_;
 };
 
 class MockServerConnection : public ServerConnection {
